refactor: remove unused existing names directive
