+++ apps/dsp-app/src/app/project/ontology/resource-class-form/resource-class-form.component.ts
@@ -6,7 +6,7 @@ import { OntologyService } from '@dasch-swiss/vre/shared/app-helper-services';
 import { OntologiesSelectors } from '@dasch-swiss/vre/shared/app-state';
 import { Store } from '@ngxs/store';
 import { Subscription } from 'rxjs';
-import { existingNamesValidator } from '../../../main/directive/existing-name/existing-name.directive';
+import { existingNamesValidator } from '../../../main/directive/existing-name/existing-names.validator';
 import { CustomRegex } from '../../../workspace/resource/values/custom-regex';
 import { atLeastOneStringRequired } from '../../reusable-project-form/at-least-one-string-required.validator';
